New Testnet Release

+++ workstation/utils.sh
@@ -8,6 +8,7 @@ REGEX_TXHASH="^[a-fA-F0-9]{64}$"
 REGEX_INTEGER="^-?[0-9]+$"
 REGEX_NUMBER="^[+-]?([0-9]*[.])?([0-9]+)?$"
 REGEX_PUBLIC_IP='^([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(?<!172\.(16|17|18|19|20|21|22|23|24|25|26|27|28|29|30|31))(?<!127)(?<!^10)(?<!^0)\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(?<!192\.168)(?<!172\.(16|17|18|19|20|21|22|23|24|25|26|27|28|29|30|31))\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(?<!\.255$)(?<!\b255.255.255.0\b)(?<!\b255.255.255.242\b)$'
+REGEX_KIRA="^(kira)[a-zA-Z0-9]{39}$"
 
 function isNullOrEmpty() {
     if [ -z "$1" ] || [ "${1,,}" == "null" ] ; then echo "true" ; else echo "false" ; fi
@@ -19,6 +20,10 @@ function delWhitespaces() {
 
 function isNullOrWhitespaces() {
     isNullOrEmpty $(delWhitespaces "$1")
+}
+
+function isKiraAddress() {
+    if ($(isNullOrEmpty "$1")) ; then echo "false" ; else [[ "$1" =~ $REGEX_KIRA ]] && echo "true" || echo "false" ; fi
 }
 
 function isTxHash() {
@@ -68,6 +73,14 @@ function isNaturalNumber() {
     if ($(isNullOrEmpty "$1")) ; then echo "false" ; else ( ($(isInteger "$1")) && [[ $1 -ge 0 ]] ) && echo "true" || echo "false" ; fi
 }
 
+function isLetters() {
+    [[ "$1" =~ [^a-zA-Z] ]] && echo "false" || echo "true"
+}
+
+function isAlphanumeric() {
+    [[ "$1" =~ [^a-zA-Z0-9] ]] && echo "false" || echo "true"
+}
+
 function isPort() {
     ( ($(isNaturalNumber $1)) && (($1 > 0)) && (($1 < 65536)) ) && echo "true" || echo "false"
 }
@@ -77,6 +90,15 @@ function isMnemonic() {
     kg_count=$(echo "$kg_mnem" | wc -w 2> /dev/null || echo -n "")
     (! $(isNaturalNumber $kg_count)) && kg_count=0
     if (( $kg_count % 4 == 0 )) && [ $kg_count -ge 12 ] ; then echo "true" ; else echo "false" ; fi
+}
+
+function date2unix() {
+    kg_date_tmp="$*" && kg_date_tmp=$(echo "$kg_date_tmp" | xargs 2> /dev/null || echo -n "")
+    if (! $(isNullOrWhitespaces "$kg_date_tmp")) && (! $(isNaturalNumber $kg_date_tmp)) ; then
+        kg_date_tmp=$(date -d "$kg_date_tmp" +"%s" 2> /dev/null || echo "0")
+    fi
+
+    ($(isNaturalNumber "$kg_date_tmp")) && echo "$kg_date_tmp" || echo "0"
 }
 
 function isPortOpen() {
@@ -181,9 +203,12 @@ function isSimpleJsonObjOrArrFile() {
 }
 
 function jsonParse() {
-    QUERY="" && INPUT=$(echo $1 | xargs 2> /dev/null 2> /dev/null || echo -n "")
-    FIN="" && [ ! -z "$2" ] && FIN=$(realpath $2 2> /dev/null || echo -n "")
-    FOUT="" && [ ! -z "$3" ] && FOUT=$(realpath $3 2> /dev/null || echo -n "")
+    local QUERY=""
+    local FIN=""
+    local FOUT=""
+    local INPUT=$(echo $1 | xargs 2> /dev/null 2> /dev/null || echo -n "")
+    [ ! -z "$2" ] && FIN=$(realpath $2 2> /dev/null || echo -n "")
+    [ ! -z "$3" ] && FOUT=$(realpath $3 2> /dev/null || echo -n "")
     if [ ! -z "$INPUT" ] ; then
         for k in ${INPUT//./ } ; do
             k=$(echo $k | xargs 2> /dev/null || echo -n "") && [ -z "$k" ] && continue
@@ -193,8 +218,8 @@ function jsonParse() {
     fi
     if [ ! -z "$FIN" ] ; then
         if [ ! -z "$FOUT" ] ; then
-            rm -f "$FOUT"
-            python3 -c "import json,sys;fin=open('$FIN',\"r\");fout=open('$FOUT',\"w\",encoding=\"utf8\");obj=json.load(fin);json.dump(obj$QUERY,fout,separators=(',',':'),ensure_ascii=False);fin.close();fout.close()"
+            [ "$FIN" != "$FOUT" ] && rm -f "$FOUT" || :
+            python3 -c "import json,sys;fin=open('$FIN',\"r\");obj=json.load(fin);fin.close();fout=open('$FOUT',\"w\",encoding=\"utf8\");json.dump(obj$QUERY,fout,separators=(',',':'),ensure_ascii=False);fout.close()"
         else
             python3 -c "import json,sys;f=open('$FIN',\"r\");obj=json.load(f);print(json.dumps(obj$QUERY,separators=(',', ':'),ensure_ascii=False).strip(' \t\n\r\"'));f.close()"
         fi
@@ -212,6 +237,7 @@ function isFileJson() {
 }
 
 function jsonQuickParse() {
+    local OUT=""
     if [ -z "$2" ] ; then
         OUT=$(cat | grep -Eo "\"$1\"[^,]*" 2> /dev/null | grep -Eo '[^:]*$' 2> /dev/null | xargs 2> /dev/null | awk '{print $1;}' 2> /dev/null 2> /dev/null)
     else
@@ -221,6 +247,67 @@ function jsonQuickParse() {
     OUT=${OUT%\}}
     ($(isNullOrEmpty "$OUT")) && return 1
     echo "$OUT"
+}
+
+function jsonEdit() {
+    local QUERY=""
+    local FIN=""
+    local FOUT=""
+    local INPUT=$(echo $1 | xargs 2> /dev/null 2> /dev/null || echo -n "")
+    local VALUE="$2"
+    [ ! -z "$3" ] && FIN=$(realpath $3 2> /dev/null || echo -n "")
+    [ ! -z "$4" ] && FOUT=$(realpath $4 2> /dev/null || echo -n "")
+    [ "${VALUE,,}" == "null" ] && VALUE="None"
+    [ "${VALUE,,}" == "true" ] && VALUE="True"
+    [ "${VALUE,,}" == "false" ] && VALUE="False"
+    if [ ! -z "$INPUT" ] ; then
+        for k in ${INPUT//./ } ; do
+            k=$(echo $k | xargs 2> /dev/null || echo -n "") && [ -z "$k" ] && continue
+            [[ "$k" =~ ^\[.*\]$ ]] && QUERY="${QUERY}${k}" && continue
+            ($(isNaturalNumber "$k")) && QUERY="${QUERY}[$k]" || QUERY="${QUERY}[\"$k\"]" 
+        done
+    fi
+    if [ ! -z "$FIN" ] ; then
+        if [ ! -z "$FOUT" ] ; then
+            [ "$FIN" != "$FOUT" ] && rm -f "$FOUT" || :
+            python3 -c "import json,sys;fin=open('$FIN',\"r\");obj=json.load(fin);obj$QUERY=$VALUE;fin.close();fout=open('$FOUT',\"w\",encoding=\"utf8\");json.dump(obj,fout,separators=(',',':'),ensure_ascii=False);fout.close()"
+        else
+            python3 -c "import json,sys;f=open('$FIN',\"r\");obj=json.load(f);obj$QUERY=$VALUE;print(json.dumps(obj,separators=(',', ':'),ensure_ascii=False).strip(' \t\n\r\"'));f.close()"
+        fi
+    else
+        cat | python3 -c "import json,sys;obj=json.load(sys.stdin);obj$QUERY=$VALUE;print(json.dumps(obj$QUERY,separators=(',', ':'),ensure_ascii=False).strip(' \t\n\r\"'));"
+    fi
+}
+
+function jsonObjEdit() {
+    local QUERY=""
+    local FVAL=""
+    local FIN=""
+    local FOUT=""
+    local INPUT=$(echo $1 | xargs 2> /dev/null 2> /dev/null || echo -n "")
+    [ ! -z "$2" ] && FVAL=$(realpath $2 2> /dev/null || echo -n "")
+    [ ! -z "$3" ] && FIN=$(realpath $3 2> /dev/null || echo -n "")
+    [ ! -z "$4" ] && FOUT=$(realpath $4 2> /dev/null || echo -n "")
+    [ "${VALUE,,}" == "null" ] && VALUE="None"
+    [ "${VALUE,,}" == "true" ] && VALUE="True"
+    [ "${VALUE,,}" == "false" ] && VALUE="False"
+    if [ ! -z "$INPUT" ] ; then
+        for k in ${INPUT//./ } ; do
+            k=$(echo $k | xargs 2> /dev/null || echo -n "") && [ -z "$k" ] && continue
+            [[ "$k" =~ ^\[.*\]$ ]] && QUERY="${QUERY}${k}" && continue
+            ($(isNaturalNumber "$k")) && QUERY="${QUERY}[$k]" || QUERY="${QUERY}[\"$k\"]" 
+        done
+    fi
+    if [ ! -z "$FIN" ] ; then
+        if [ ! -z "$FOUT" ] ; then
+            [ "$FIN" != "$FOUT" ] && rm -f "$FOUT" || :
+            python3 -c "import json,sys;fin=open('$FIN',\"r\");fin2=open('$FVAL',\"r\");obj2=json.load(fin2);obj=json.load(fin);obj$QUERY=obj2;fin.close();fout=open('$FOUT',\"w\",encoding=\"utf8\");json.dump(obj,fout,separators=(',',':'),ensure_ascii=False);fin2.close();fout.close()" || SUCCESS="false"
+        else
+            python3 -c "import json,sys;f=open('$FIN',\"r\");fin2=open('$FVAL',\"r\");obj2=json.load(fin2);obj=json.load(f);obj$QUERY=obj2;print(json.dumps(obj,separators=(',', ':'),ensure_ascii=False).strip(' \t\n\r\"'));f.close();fin2.close()"
+        fi
+    else
+        cat | python3 -c "import json,sys;obj=json.load(sys.stdin);fin2=open('$FVAL',\"r\");obj2=json.load(fin2);obj$QUERY=obj2;print(json.dumps(obj$QUERY,separators=(',', ':'),ensure_ascii=False).strip(' \t\n\r\"'));fin2.close()"
+    fi
 }
 
 # e.g. urlExists "18.168.78.192:11000/download/peers.txt"
@@ -233,7 +320,7 @@ function urlExists() {
 # TODO: Investigate 0 output
 # urlContentLength 18.168.78.192:11000/download/snapshot.zip 
 function urlContentLength() {
-    VAL=$(curl --fail $1 --dump-header /dev/fd/1 --silent 2> /dev/null | grep -i Content-Length -m 1 2> /dev/null | awk '{print $2}' 2> /dev/null || echo -n "")
+    local VAL=$(curl --fail $1 --dump-header /dev/fd/1 --silent 2> /dev/null | grep -i Content-Length -m 1 2> /dev/null | awk '{print $2}' 2> /dev/null || echo -n "")
     # remove invisible whitespace characters
     VAL=$(echo ${VAL%$'\r'})
     (! $(isNaturalNumber $VAL)) && VAL=0
@@ -253,19 +340,20 @@ function globFile() {
 }
 
 function globGet() {
-    kg_FIL=$(globFile "$1" "$2")
+    local kg_FIL=$(globFile "$1" "$2")
     [[ -s $kg_FIL ]] && cat $kg_FIL || echo ""
     return 0
 }
 
 # threadsafe global get
 function globGetTS() {
-    kg_FIL=$(globFile "$1" "$2")
+    local kg_FIL=$(globFile "$1" "$2")
     [[ -s "$kg_FIL" ]] && sem --id $1 "cat $kg_FIL" || echo ""
     return 0
 }
 
 function globSet() {
+    local kg_FIL=""
     [ ! -z "$3" ] && kg_FIL=$(globFile "$1" "$3") || kg_FIL=$(globFile "$1")
     touch "$kg_FIL.tmp"
     [ ! -z ${2+x} ] && echo "$2" > "$kg_FIL.tmp" || cat > "$kg_FIL.tmp"
@@ -274,6 +362,7 @@ function globSet() {
 
 # threadsafe global set
 function globSetTS() {
+    local kg_FIL=""
     [ ! -z "$3" ] && kg_FIL=$(globFile "$1" "$3") || kg_FIL=$(globFile "$1")
     touch "$kg_FIL"
     [ ! -z ${2+x} ] &&  sem --id $kg_NAM "echo $2 > $kg_FIL" || sem --id $kg_NAM --pipe "cat > $kg_FIL"
@@ -350,8 +439,7 @@ function timerDel() {
 }
 
 function prettyTime {
-  local T=$1
-  (! $(isNaturalNumber $T)) && T=0
+  local T=$(date2unix "$1")
   local D=$((T/60/60/24))
   local H=$((T/60/60%24))
   local M=$((T/60%60))
@@ -365,18 +453,33 @@ function prettyTime {
   (( $S != 1 )) && printf '%d seconds\n' $S || printf '%d second\n' $S
 }
 
+function prettyTimeSlim {
+  local T=$(date2unix "$1")
+  local D=$((T/60/60/24))
+  local H=$((T/60/60%24))
+  local M=$((T/60%60))
+  local S=$((T%60))
+  (( $D > 0 )) && (( $D > 1 )) && printf '%dd ' $D
+  (( $D > 0 )) && (( $D < 2 )) && printf '%dd ' $D
+  (( $H > 0 )) && (( $H > 1 )) && printf '%dh ' $H
+  (( $H > 0 )) && (( $H < 2 )) && printf '%dh ' $H
+  (( $M > 0 )) && (( $M > 1 )) && printf '%dm ' $M
+  (( $M > 0 )) && (( $M < 2 )) && printf '%dm ' $M
+  (( $S != 1 )) && printf '%ds\n' $S || printf '%ds\n' $S
+}
+
 function resolveDNS {
     if ($(isIp "$1")) ; then
         echo "$1"
     else
-        kg_dns=$(timeout 10 dig +short "$1" 2> /dev/null || echo -e "")
+        local kg_dns=$(timeout 10 dig +short "$1" 2> /dev/null || echo -e "")
         ($(isIp $kg_dns)) && echo $kg_dns || echo -e ""
     fi
 }
 
 function isSubStr {
-    STR=$1
-    SUB=$2
+    local STR=$1
+    local SUB=$2
     [[ $STR == *"$SUB"* ]] && echo "true" || echo "false"
 }
 
@@ -385,34 +488,54 @@ function isCommand {
 }
 
 function isServiceActive {
-    ISACT=$(systemctl is-active "$1" 2> /dev/null || echo "inactive")
+    local ISACT=$(systemctl is-active "$1" 2> /dev/null || echo "inactive")
     [ "${ISACT,,}" == "active" ] && echo "true" || echo "false"
 }
 
 # returns 0 if failure, otherwise natural number in microseconds
 function pingTime() {
     if ($(isDnsOrIp "$1")) ; then
-        PAVG=$(ping -qc1 "$1" 2>&1 | awk -F'/' 'END{ print (/^rtt/? $5:"FAIL") }' 2> /dev/null || echo -n "")
+        local PAVG=$(ping -qc1 "$1" 2>&1 | awk -F'/' 'END{ print (/^rtt/? $5:"FAIL") }' 2> /dev/null || echo -n "")
         if ($(isNumber $PAVG)) ; then
-            PAVGUS=$(echo "scale=3; ( $PAVG * 1000 )" | bc 2> /dev/null || echo -n "")
+            local PAVGUS=$(echo "scale=3; ( $PAVG * 1000 )" | bc 2> /dev/null || echo -n "")
             PAVGUS=$(echo "scale=0; ( $PAVGUS / 1 ) " | bc 2> /dev/null || echo -n "")
             ($(isNaturalNumber $PAVGUS)) && echo "$PAVGUS" || echo "0"
         else echo "0" ; fi
     else echo "0" ; fi
 }
 
-displayAlign() {
-  align=$1
-  width=$2
-  text=$3
+function pressToContinue {
+    if ($(isNullOrEmpty "$1")) ; then
+        read -n 1 -s 
+        globSet OPTION ""
+    else
+        while : ; do
+            local kg_OPTION=""
+            local FOUND=false
+            read -n 1 -s kg_OPTION
+            kg_OPTION="${kg_OPTION,,}"
+            for kg_var in "$@" ; do
+                kg_var=$(echo "$kg_var" | tr -d '\011\012\013\014\015\040' 2>/dev/null || echo -n "")
+                [ "${kg_var,,}" == "$kg_OPTION" ] && globSet OPTION "$kg_OPTION" && FOUND=true && break
+            done
+            [ "$FOUND" == "true" ] && break
+        done
+    fi
+    echo ""
+}
 
-  if [ $align == "center" ]; then
-    textRight=$(((${#text} + $width) / 2))
-    printf "|%*s %*s\n" $textRight "$text" $(($width - $textRight)) "|"
-  elif [ $align == "left" ]; then
-    textRight=$width
-    printf "|%-*s|\n" $textRight "$text"
-  fi
+displayAlign() {
+    local align=$1
+    local width=$2
+    local text=$3
+
+    if [ $align == "center" ]; then
+        local textRight=$(((${#text} + $width) / 2))
+        printf "|%*s %*s\n" $textRight "$text" $(($width - $textRight)) "|"
+    elif [ $align == "left" ]; then
+        local textRight=$width
+        printf "|%-*s|\n" $textRight "$text"
+    fi
 }
 
 function echoInfo() {
